[win] fix cross-compilation for arm64
Cross-compiling x64 to ARM64 on Windows fails with both clang and MSVC.
The reason is the CONTEXT structure which does not have the Pc field.
Using only CONTEXT.Rip fixes the compilation issue.

This follows https://chromium-review.googlesource.com/c/v8/v8/+/4403215
as part of an effort to completely fix cross-compilation with MSVC.

Change-Id: I2148349b316fb6f58e66ae69a3ef081624f4b422