[wasm] Call JS-linkage builtins by JSFunction
Instead of calling JS-linkage builtins by their builtin ID, we now
store the originally-imported JSFunction on the instance, and use
it for the call.
This allows us to maintain the invariant that stack frames of such
builtins have an actual JSFunction in their "function" slot.

This patch reverts some of the new infrastructure recently introduced
in crrev.com/c/4503230, in favor of a more robust mechanism.

Fixed: v8:13985
Change-Id: If6e9a21c6664af952fa8d86a18b77c069d7f74ba