Implement a debugging extension for lldb
The debug_helper library is meant to be cross-platform and
debugger-agnostic, but is currently only used by v8windbg. This change
introduces a new debugging extension for lldb that includes a subset of
the same features as v8windbg, in the hope that it might be useful. This
extension adjusts the behavior when printing V8 Objects and Handles with
`frame variable` or its various shorthand forms, so that the user can
easily inspect the fields of V8 heap objects. An example interaction is
pasted below, lightly reformatted for width. In this example, `object`
is a local variable of type v8::Internal::Handle<v8::Internal::Object>.

(lldb) frame variable -D 1 object.Value.map.instance_descriptors
(v8::internal::TaggedValue) object.Value.map.instance_descriptors =
    0x1c510004c665 <v8::internal::DescriptorArray> {
  ptr_ = 312933
  map = DescriptorArrayMap (0x1c5100000129 <v8::internal::Map>){...}
  number_of_all_descriptors = 4
  number_of_descriptors = 4
  raw_gc_state = 0
  enum_cache =
      EmptyEnumCache (0x1c5100000289 <v8::internal::EnumCache>){...}
  descriptors[0].key =
      "a" (0x1c5100002a55 <v8::internal::SeqOneByteString>){...}
  descriptors[0].details = 1602 (0x642){...}
  descriptors[0].value = 1 (0x1){...}
  descriptors[1].key =
      "b" (0x1c5100002a65 <v8::internal::SeqOneByteString>){...}
  descriptors[1].details = 525506 (0x804c2){...}
  descriptors[1].value = 1 (0x1){...}
  descriptors[2].key =
      "c" (0x1c5100002a75 <v8::internal::SeqOneByteString>){...}
  descriptors[2].details = 1049282 (0x1002c2){...}
  descriptors[2].value =
      weak ref to 0x1c51001881c9 <v8::internal::Map>{...}
  descriptors[3].key = "d"
      (0x1c5100002a85 <v8::internal::SeqOneByteString>){...}
  descriptors[3].details = 1572994 (0x180082){...}
  descriptors[3].value = 1 (0x1){...}
}

Bug: v8:9376
Change-Id: Ia0c5a9a870976b723c6bf373b91c95eeb4843abe