[test] Move convenience status-file variables into BUILD.gn
This creates the composed properties used for convenience already
in BUILD.gn and drops logic need for those in the test framework.
Maintaining build-flags from the build config is now much simpler
in the test framework.

This also:
- Adds a meaningful exception and early checking to all places
where a build flag might be missed. Previously it was confusing
to find the right place where to add it.
- Refactors _no_simd_hardware (leftover from last review).
- Removes unused build flags like mips_arch and write_barriers,
only used in other compounds.
- Uses positive expression for no_js_shared_memory ->
not js_shared_memory, as in all other flags.

Bug: chromium:1132088
Change-Id: Iedeb8c54a396bd099e9e8b22bece5ac10cd63ea9