[heap] Introduce witness objects for parking/unparking
References to ParkedScope objects are often passed as parameters to
functions that expect the current thread to be parked. This CL refactors
this mechanism and introduces explicit witness objects for this purpose.
Witness objects have no run-time effect in non-debug builds, whereas in
debug builds they provide an improved and more generic mechanism for
run-time checks.

The CL also refactors the classes in parked-scope.h that are used for
waiting on resources while the thread is parked.

Bug: v8:13257
Change-Id: I41adfc269640be56da05af9d55e1b3ad964d4a0a