[wasm] Avoid creating temporary SmallVectors
{PopArgs} currently returns a {base::SmallVector} containing the popped
values. As Liftoff does not use those values, we should avoid creating
this copy.

This CL implements this by specifying {kUsesPoppedArgs} per interface,
and making {PopArgs} return an empty data structure if it is false.
This requires a few changes to interface methods to not expect a
{base::Vector} but take a direct pointer instead. If {kUsesPoppedArgs}
is {false}, this pointer will always be a {nullptr} then.

R=jkummerow@chromium.org

Bug: v8:13976
Change-Id: I262bc9ba329bad1a4cd8f8e9c0915a5e4a65ae15