[wasm] Fix signatures of atomic opcodes
Decoding of atomic opcodes happens via their static signature. This
signature currently encodes an 'i32' first parameter for the memory
index, even though this is either i32 or i64, depending on the declared
memory type.

This CL thus fixes the static signatures to use the bottom type for the
first parameter. This allows us to use the standard {PopArgs} method for
popping the input values. The memory index is then validated separately,
depending on the memory type.

R=ahaas@chromium.org

Bug: v8:13976
Change-Id: I4247d9d69507a82ca5712483e1bd1d91f273d262