[wasm][multi-memory] Store memories in a vector
This CL prepares support for multiple memories. Instead of storing
informaiton about the one memory in the {WasmModule} itself, we store a
vector of {WasmMemory} objects.

For now, this vector is limited to a size of 1, so there should be no
observable change in behaviour (except for some slightly different error
messages).
A follow-up CL will introduce a flag to allow for more than one
memory and add some tests for that.

R=thibaudm@chromium.org

Bug: v8:13918
Change-Id: I25b3ab28d68cafc70ed18d097bc430ca91ec0183