[wip][arm64] Optimise veneering for long link chains
Applying veneers to out of range branches when the link chain is long
can be extremely slow. Profiling reveals >95% of the time is spent
walking the link chain.

Making the chain doubly-linked accelerates removing a link from the chain significantly. The back edges can't be stored in the code itself
(where the forward edges live) so maintain an unordered map in the
assembler.

Bug: chromium:1427422
Change-Id: Ic98d6bea0f0fffebb8a451479b9e308ec62d3ced