[builtins] Refactor add logic in SetPrototypeAdd and MapPrototypeSet
This CL refactors common logic to add/set an elements in an
OrderedHashTable out of implementations of SetPrototypeAdd and
MapPrototypeSet into a separate, templated method.

This is a prerequisite CL for implementing Set methods and the static
method version of Array grouping. Both these features need to add/set
elements (in OrderedHashSet and OrderedHashMap, respectively).

CollectionsBuiltinsAssembler is also moved to the header file so it can
be used from Torque.

Bug: v8:13556, v8:12499
Change-Id: Id09c691e399ce77562ea2af63bbd20bc300f843d