[maglev] Bail out early on empty map checks
Emit an unconditional deopt for impossible map checks (i.e. when we're
checking an empty set of maps, which happens when known map inference
intersects with feedback to create an empty set).

Similarly, InferHasInPrototypeChain can get an empty set of maps, which
should also bail out early (as "definitely not in chain").

Bug: v8:7700
Change-Id: I83543f344913d3ba94d0be30f3994599941c2c83
Fixed: chromium:1438990