[maglev] OSR-in
OSR-in for maglev. Not yet enabled.

When `--maglev --maglev-osr` is given then OSR will first tier up to
Maglev. The flag also weakly implies `--osr-from-maglev` meaning we
can further tier up to TF. This happens only for OSR'd code. The idea
being that if we didn't OSR into Maglev in the first place chances are
slim that we are really stuck and OSR to TF is needed.

Change-Id: I8c721a0c95f15740ef350b5af6bbf0e8c98719ff