Stop sampling even if no more JS runs
TracingCpuProfilerImpl::OnTraceDisabled requests an interrupt to stop
profiling, but if the associated Isolate doesn't run any more script, it
never runs that interrupt and thus leaves the sampler thread active.
This change updates OnTraceDisabled to also post a task which will run
any outstanding interrupts, so that the sampler doesn't keep spinning
indefinitely.

Bug: chromium:1448583
Change-Id: I1c6cbd2995c0cb08a1c8f71448178a8f36acad86