[turbofan] Use start as the control input of Projection
For each node in TurboFan, at most one Projection node is allowed per
output. Therefore, if GVN merges two nodes with Projection nodes as
users, also the Projection nodes have to be merged. However, if the
Projection nodes have different control inputs, they cannot be merged.

In principle The Projection node does not need a control input. However,
the control input enforces better scheduler decisions when the
projection is used for deopt decisions.

This CL therefore does not remove the control input of the Projection
node, but uses the start node as the control input for Projection nodes
created by the GraphAssembler. The GraphAssembler is not used for code
dealing with deopts, but it is used by the fast API calls where the
problems with Projection node merging showed up.

R=dmercadier@chromium.org

Bug: chromium:1449115
Change-Id: I84ea596762283e11a38947c59948bd4bc5ca7184