[pku] split ThreadIsolationData into trusted and untrusted
Signal handlers run without read access to non-default pkey memory.
Once we start pkey-tagging the code space, we need to allow read access
to the pkey. However, we also store the pkey in pkey-tagged memory.

To support this, make a copy of the pkey and store it in untrusted
memory and also move the debug-only data in there.

Change-Id: I76e9cf8293848fc4efb91d457ca6f8f7db396a69
Bug: v8:13355