[tools] Add iOS command execution through iOS Simulator
This CL adds the ability to test runners on iOS for mjsunit (using d8).
Rather than having to use an actual device (iPhone, iPad, etc), we use
the iOS Simulator for the test runners and in order to ease the job of
builders and testers that won't require a physical device.

At the moment Chromium's iossim tool is being used, which is a wrapper
around 'simctl' macOS command utility.

I've tested with variations of the following command line:
./tools/run-tests.py --progress=verbose --outdir ../out/Debug-iphonesimulator "mjsunit/delete*"

Bug: v8:13972
Change-Id: Ibe18b0da75cf3b452c507d1a05c88b8aa473ed54