[ic] Fix clone ic creating detached maps
Currently the clone ic creates detached maps. When these maps transition
we always create new maps leading to pretty bad performance cliffs.

This CL fixes the ic by not duplicating the map anymore. This slightly
reduces the applicable cases (see TODOs). However, we get rid of the
performance cliffs.

In particular the clone ic fast case applies when (a) the result map
can be identical to the source map. Or, when (b) the result map has
only maximally generic representations, such that an unrelated source
map representation change cannot invalidate it.

The slow (i.e. megamorphic) case of the IC, or clones falling under
(b) use the object literal map cache as the root map.

Based on the prototype in
 https://crrev.com/c/3045338

Bug: chromium:1204540
Change-Id: I5166c111274da86eaad8b102e3ec2f7c1520934f