[snapshot] Always use memcpy-based serialization for RO space
V8's snapshot now has two serialization methods: 1) serialization
through depth-first object iteration, and 2) serialization through
linear space iteration in which objects are visited in the same order
as they occur in memory.

Most serializers, notably for the isolate- and context snapshots,
still use 1). The read-only snapshot was switched to 2) for
static-roots builds a while ago, and this CL switches it over for
all remaining build configs as well.

The main motivation is that code caches depend on a specific RO
space layout, and thus RO space layout must not change when code
caching is involved. Method 2) guarantees this.

As a side benefit, this again unifies RO space serialization
mechanisms, and gets rid of the ro object cache.

Memcpy-based deserialization:

1. copies the raw memory layout over in a first pass,
2. relocates tagged pointers in a second pass (if needed, see
   kRelocateSegment), and
3. does final custom patchups (e.g. setting Code::instruction_start) in
   a final pass (PostProcessNewObjects).

Bug: v8:13789
Change-Id: I1f69049d7b001db04c8003a9c47727abe0b0ece9