[turbofan] Fix String builder issue with multiple loop exits
We previously assumed that loop had a single exit. However, this is
wrong and can lead to some paths not having the trimming required for
the end of the string builder. This CL fixes that.

However, the string builder is still broken, because the trimming
produced on those multiple loop exists could need merging before being
used, which requires computing where to merge and then inserting new
phis. This is a bit hard to do in Turbofan and will be easier to do in
Turbosahft (thanks to the VariableReducer), so we'll probably not land
this CL and wait for the string builder to be ported to Turboshaft.

Bug: chromium:1449589
Change-Id: I1851896aca094bee3c3a050f43ad27c0bffb38df