[wasm-gc] Inlining into JS: Lower traps to conditional jump to trap call
See bug v8:14013, the experimental inlining can cause crashes due to
stale references on the JS stack frame from which an inlined wasm trap
was called. The value can get stale if a GC is triggered during
creation of the JS error representing the wasm trap or during
creation of the stack frame summary in OptimizedFrame::Summarize.

This is not an issue for wasm as traps are uncatchable from within
wasm and the values on the top stack frame are not inspected at stack
trace creation.

Fixing this is non-trivial as we do not only have to track references
in the safepoint for the wasm trap, we also might need to generate
spill instructions in the out of line code of the trap.
As the out of line code is generated after instruction selection,
this is not directly possible.

Solution: In the case of wasm inlined into JS replace all wasm trap
nodes (at a late point in the pipeline to still use all optimization
code for them) with a conditional branch to out of line code calling
the corresponding trap builtin.
All the safepoint, deopt point and spill instructions creation is then
automatically taken care of by the compiler infrastructure as it looks
like any other builtin call in JavaScript.

Note: Due to doing this replacement late in the pipeline, the reducer
has to be reimplemented in TurboShaft as well. This is a temporary
duplication that will disappear once TurboShaft is enabled in
production and support for the non-turboshaft code path can be dropped.

Bug: v8:14013, v8:7748
Change-Id: I9203a9a19d5f870ac130e5df91b62dbffb220e92