[wasm-gc] Inlining into JS: Get rid of SourcePositionTable at runtime
The SourcePositionTable is still needed during compilation and is
enabled lazily for any JS function containing a potentially inlineable
wasm call if the experimental inlining is enabled.

The runtime SourcePositionTable is not created (if not required e.g.
by profiling or debugging). To make this possible, the source position
information for an trap location is stored in the FrameState similar
to how it is done for JavaScript exception locations. (The offset is
however the wirebytes offset and not the bytecode offset.)

Bug: v8:7748
Change-Id: I310cba3540a001872c239270e067169fdd9d5cff