[heap] Stop checking marking worlists for main threads of clients
When allocating in the shared space from a client isolate, the
ConcurrentAllocator invokes Heap::ShouldExpandOldGenerationOnSlowAllocation and then Heap::IsMajorMarkingComplete for the heap of the shared space isolate.
IsMajorMarkingComplete checks the local marking worklist of the
main thread of the shared space isolate which is not thread-safe.
Hence checking it from a client isolate's main thread is illegal.
So far we already bail out for background threads. This CL
now also bails out for client isolate main threads.

Bug: chromium:1450809
Change-Id: Ib468979d61a8d3a99cd9547d48e7587df9956459