[maglev] Avoid emitting phis for exception context
When the handler context register has a known value, we can avoid
emitting a Phi for the value, by using the same merge logic on the
handler-specific context register as we do for other registers. In
practice, this means that the current context is always statically
known, except when there is a generator.

Bug: v8:7700
Change-Id: Ib00466f95ef582ef97bdff3ca44db791d77d024b
Fixed: chromium:1450594