[x64] Enable jump table switch for builtin
In V8 compiler, there are two methods to implement a 'switch' statement:
Table Switch and Binary Search Switch. If the amount of cases is large
and the values are contiguous, the Table Switch is usually better
choice.

Currently, the jump table will store absolute address of the jump
target, it's position dependent, so it can not be used in builtins
functions because their code positions may be changed.

To solve this problem, when compile builtins, we replace the values
stored in the table from `target_address` to
`target_address-table_address` and it will become position independent,
then we can enable Table Switch for builtins as well.

As for the disassembler for such switch table, this CL add a new
RelocInfo mode: "RELATIVE_SWITCH_TABLE_ENTRY" for recording the position
of the table entries and the postion of the jump target.

Bug: v8:13710

Change-Id: I2f2ef10d739f9e4419e506ae0e4c164d0cf51309