[sfi] Unconditionally enable the SFI::unique_id field
The motivation is that a unique SFI id can be used as the key in a
plain old unordered_map (and thus lets us avoid problematic
IdentityMap uses).

In kTaggedSize == 4 configs, the SFI grows by 4 bytes. In kTaggedSize
== 8 configs, the SFI size remains unchanged (due to removed padding).

Note this is in preparation for removing ~800 SFIs from the
per-Isolate snapshot, so the size increase should be more-than offset
in the common case.

Bug: v8:13789
Change-Id: I523e9aa3fed8604555ec0c1d821e44460d16fe19