[maglev] Disable branch fusion for inlined constructors
Constructor calls replace returned primitives with the new.target, so
the cross-inlining branch fusion has to be aware of this. In practice,
the easiest way to fix this is to disable cross-inlining branch fusion
when returning from a constructor, i.e. when a new.target is present.

As a side effect, we are now passing in the correct new.target to
inlined functions, which means we can enable inlining them in the near
future.

Bug: v8:7700
Change-Id: Ib21b5faf9ea057a3ffc1885fae9c2e36c1e1be1c
Fixed: chromium:1454478