[cleanup][arm] Remove kR9Available
This has been in the codebase for at least 15 years:
https://github.com/v8/v8/commit/43d26ecc3563a46f62a0224030667c8f8f3f6ceb

If there was a platform using it as described in the comment,
we would probably have found by now.

If such platform appears in the future, there is no guarantee
that the code would work with hR9Available=0, since this has never
been tested.

Change-Id: I9c8cd1dbcf70400495848b2a97fdd14c2f049deb