[wasm] Allocate instance fields before the instance itself
We currently allocate the {WasmInstanceObject} and then allocate
individual objects (FixedArrays, ByteArray, ...) for the fields of the
instance.
When allocating the arrays for the fields, the instance object itself is
not fully initialized, which can lead to heap verification issues.
Thus do first allocate all objects stored in the instance fields, and
then allocate the instance.

R=thibaudm@chromium.org
CC=dlehmann@chrmium.org

Bug: v8:13918
Change-Id: Ifc44aad65dd669919d4ff3699bb67c568033a2db