[foozzie] Decode output after capping
Before this fix, we'd first decode the output of each run to compare
and afterwards cap the longer output if one had a crash and is
shorter. The problem is that the longer output might have characters
only in the excess part that lead to decode errors, using a different
encoding on the whole output. This in turn might change characters
in the shared part.

We can't trust any output after one of the runs has crashed and
already don't consider it for difference comparison. In should
also not lead to any encoding changes of the parts we trust,
which is fixed now by delaying the decoding of bytes until after
we cap their length of both comparison runs.

If there is an encoding difference in the remaining shared part
of the output that we compare, it would point to an actual
logical difference and a real bug, which we'd still detect.

This CL retains Py2/Py3 compatibility, which is probably not
necessary anymore and will be dealt with in a separate CL.

Bug: chromium:1448728
Change-Id: I0354ed4c0b26ccc34275dab28dbf58efbeaa16fc