[turboshaft] Constant-fold heap object map loads when possible
... and also constant-fold Smi-checks on Heap Objects constants.

Drive-by: fix a bug with GOTO_IF and GOTO_IF_NOT when the destination
was not reachable through this specific GOTO_IF: in such cases, we
would record a Phi input which we shouldn't have recorded.

Drive-by: when lowering/optimizing a DeoptimizeIf to Deoptmize,
generating_uneachable_operations_ wasn't set, causing a DCHECK failure
on the next generation of the operations since current_block_ was
nullptr. This is now fixed: DeoptimizeIf(Not) and TrapIf(Not) properly
set generating_uneachable_operations_ when needed.

Bug: v8:12783
Change-Id: I12f9044e5449341302665759238e71086b284ce5