[builtins] Make GetRegisterParameterCount aware of FP registers
GetRegisterParameterCount counts the number of valid registers in the
registers array defined for an InterfaceDescriptor. With the addition
of FP registers the register array may contain `no_reg` where the
signature requires an FP register and not a GP register. Therefore
GetRegisterParameterCount counts now the number of consecutive array
indices starting at index '0' which contain either a valid GP register
or a valid FP register.

R=nicohartmann@chromium.org

Bug: v8:14035
Change-Id: If177322263de43922712020b3d536d628c9e8925