[maglev] Ensure every deopt frame has valid source position
Initialize the current_source_position_ to a valid position given
entrypoint_ early in graph building. This ensures the initial stack
check deopt frame contains a valid position.

The reason we need this is when this initial deopt frame gets used for
eager deopts (e.g. in the case of deopt for osr-from-maglev), then the
profiler expects it to have a correct source position.

Drive-By: Cleanup and register alias check in
AttemptOnStackReplacement

Bug: chromium:1452126
Change-Id: I62e5bd532e125b89b3e2f6bc318d02022914d6c2