[turboshaft] Constant-fold heap object map loads when possible
... and also constant-fold Smi-checks on Heap Objects constants.

Drive-by: fix a bug with GOTO_IF and GOTO_IF_NOT when the destination
was not reachable through this specific GOTO_IF: in such cases, we
would record a Phi input which we shouldn't have recorded.

Drive-by: when lowering/optimizing a DeoptimizeIf to Deoptmize,
generating_uneachable_operations_ wasn't set, causing a DCHECK failure
on the next generation of the operations since current_block_ was
nullptr. This is now fixed: DeoptimizeIf(Not) and TrapIf(Not) properly
set generating_uneachable_operations_ when needed.

Drive-by: some MachineLoweringReducer code was using
`IF .. ELSE .. END_IF` followed by manually-created Phis, which
doesn't work well for cases where the IF's condition is optimized
to a Constant (since the Phi now has a single predecessor instead of
two). I've switched to using Variables instead (since they
automatically insert Phis with the correct inputs).

Drive-by: when generating unreachable operations, a DCHECK was
failing AllocateBigInt. I've made sure that we don't call
AllocateBigInt when we're generating unreachable operations.


Bug: v8:12783
Change-Id: I12f9044e5449341302665759238e71086b284ce5