[regalloc] Add datastructure for use positions and intervals
The linked lists of `UsePosition`s and `UseInterval`s in each
`LiveRange` are highly cache unfriendly for pathological functions,
see the bug below.
We can address the performance issues by storing them contiguously
instead, but this introduces high max zone memory usage if not careful.

This CL introduces a special-purpose datastructure, see the class
comment for the design goals.

Bug: v8:14082
Change-Id: I6c0c529a98ca5841df0c23bf7b9c8b86b9c4b178