[regalloc] Efficient insert of UsePositions at front
Use the just added `DoubleEndedSplitVector` datastructure, which allows
efficient inserts at the front, as is (mostly) the case in practice
for `UsePosition`s.

This avoids reallocations and copies in pathological cases, e.g.,
- function #52 of 401.bzip2 benchmark (SPEC CPU 2006):
  reduces total compile time by ~47% (-118ms)
- function #92 of HashSet JetStream 2 benchmark:
  reduces total compile time by ~43% (-38ms)
- function #2839 of binary from v8:9529:
  reduces total compile time by ~3% (-260ms)
No regressions in terms of total or maximum Zone memory usage,
or average, median, or 99% percentile total compile time for all
Wasm functions in our benchmarks.

Bug: v8:14082, v8:9529
Change-Id: Ia7cf78ec61accccee28d1798928db760262cac69