[regalloc][cleanup] Bundles belong to TopLevelLiveRanges
The `bundle_` for all child `LiveRange`s (which are created by
splitting and merging again) of the same `TopLevelLiveRange` are always
the same, so we can move the field to the `TopLevelLiveRange`.

This doesn't change behavior (and I couldn't measure a difference in
compile time), but reduces max Zone memory usage by ~0.5%, especially
for pathological examples where there are many child live ranges.

Bug: v8:14082
Change-Id: I0a722434ee35fdf1a65e0c80cc933b11968936d2