[regalloc] Linked list of UseIntervals to vector
`UseInterval`s are used in `LiveRange`s, `LiveRangeBundle`s and
`SpillRange`s for intersection tests. So far, they were stored in a
linked list, which was cache unfriendly for pathological cases
(see bugs below) with ~50k intervals for some live ranges.

This CL changes those linked lists (in particular in `LiveRange`, but
also in `SpillRange`) to a `DoubleEndedSplitVector`, which has exactly
the properties for efficient insertion, splitting, and merging that we
need here
(see https://crrev.com/c/4650483/3).

It improves total compile time for those pathological cases:
- function #65374 in the Photoshop Web binary in our benchmarks:
    -60% (900ms)
- function #2839 in the binary of v8:9529: -18% (1240ms)
- function #1407 in the binary v8:12320: -9% (760ms)

No signficant regressions in mean, median, 90th, or 99th percentile
compile times. A slight maximum Zone Memory usage regression
(1.3% or 1.5KB for the median function, 2.4% or 58KB for the 99th
percentile, 1.7% or 13.5MB in the worst case), but I think that's
justified and should not affect JavaScript (which has much smaller
functions than those pathological examples with ~2MB bytecode size).

Bug: v8:14082, v8:9529, v8:11237, v8:12320, chromium:974804
Change-Id: I3a74d64ce1048ca9927c91f19507714a0b1b55fe