[turbofan] Add limit on number of tracked fields in LoadElimination
LoadElimination already had a limit on the number of tracked objects
and the number of tracked fields per objects, both aiming at providing
a not-so-bad worst-case execution time. However, since this pass is
more or less quadratic in the worst case, this is still not enough.
The number on the number of tracked fields should help get the
worst-case performance under control.

On a JS program that does 1000 assignments, like:

    let r = {}
    r["a0"] = ...
    r["a1"] = ...
    ....
    r["a999"] = ...

LoadElimination takes now 220ms instead of 3200ms. This is still
probably too much, but the whole compile time is around 4.5sec, so
that's only 6% in LoadElimination, which is probably fine.

Note that for this specific example, a better fix would be to have 2
indices in LoadElimination: one on the offsets and one on the
objects. The former would provide fast lookup of all of the objects
that have known data at a specific field. The latter would provide
fast lookup of all of the fields of an object. LoadElimination
currently only has the former, and having the latter would
significantly improve performance.

Given that we're planning to replace Turbofan by Turboshaft in the
near future, it's probably not worth spending time adding this second
index in Turbofan: we'll just get it right in Turboshaft from the
start. (and also, Turboshaft won't require copying the state all
around, so this should speed up performance as well)

Change-Id: Iffff99e31a51533a898d8fa0747fa737d6f4d2b1