[heap] Reset global limit on global context disposal
Global context disposal (no dependent contexts) resets the regular JS
heap limit and marks the old generation as not configured. This is
used for embedders to signal that all contexts went away and a new
workload is started.

Also reset the global heap limit in this disposal as the JS and C++
heaps are intertwined.

Sheriffs: This may move GC scheduling a lot as it likely delays the
first GC in benchmarks. It's possible to see average memory
consumption go up and GC time decrease.  Steady-state memory should be
unaffected though.

Bug: v8:13012
Change-Id: I2036d5a8a5f147bcfbcb92c35f9f8b7b42f070ad