[maglev] Request install after every maglev job execution
With a a maximum maglev concurrency of one, we may end up batching a
bunch of maglev compilations and not finalize them until they all
complete. Instead of this, request a finalization interrupt on ever
executed job -- it may cause us to interrupt more, but for that we'll
get to run more maglev code earlier so it's likely worth it.

Also, interleave popping execution and destruction tasks in a single
loop on the task, so that if a compile job is enqueued while
destructions are happening, we switch back to compiling.

Bug: v8:7700
Change-Id: Idefa6b4292609f4c9210b3d54f7c6e70342fe979