[inspector] Hold on weakly to the evaluation promise
This CL fixes a bug where we hold on strongly to the promise of
evaluations in the ProtocolPromiseHandler.
This is not a problem in most cases since evaluations
either resolve or an object group keeps them alive.

Nonetheless, we should not tempt fate and only hold on weakly.

Drive-by: As a small cleanup we now pass the resolver promise to the
ProtocolPromiseHandler in cases where the evaluation didn't result in
a promise itself.

R=jarin@chromium.org

Bug: None
Change-Id: I798394bb70f7db5045e794fcb89fd4aa49d441ab