[heap] Introduce separate visitor for reference summarizer
This CL implements a MarkingVisitor for the ReferenceSummary. So far
it was re-using MainMarkingVisitor but that also means that it would
reuse the implementation of e.g. MainMarkingVisitor::RecordSlot which
might record a OLD_TO_OLD slot when no GC is active. I believe it's
not an issue yet because we don't have evacuation candidates at that
point. However, it's probably cleaner for that reason to have a
separate visitor.

Bug: v8:14135
Change-Id: Ic2d77d794b3564e7e6bf88d33243dac028856e03