[maglev-osr] Add a maybe_has_maglev_osr_code flag to the fv
Sharing the maybe_has_optimized_osr_code flag on the feedback vector for
maglev and turbofan osr caused the osr checks in osr-from-maglev to
trigger unneccessarily often, since the flag was already set by
maglev-osr.

With this change we add an additional flag to distinguish the two cases.
For interpreter and baseline code nothing changes as these tiers will
happily tier up to either maglev or turbofan.

Bug: v8:7700
Change-Id: I89d5ad0eec7f7b53a7a8850f48d563d179324850