[object] Add concrete subtypes for the oddballs
Move the concept of separate types for Undefined/Null/True/False from
just being in torque into the C++ object graph. This includes adding a
Boolean subtype of Oddball, which take the ToBool method (this is anyway
DCHECKed to be only called for Boolean)

This has the consequence that True and False are separate types, and
thus Handle<True> and Handle<False> don't have an obvious, to C++,
inferrable conversion to Handle<Boolean> (e.g. in ternary operators), so
some casting is necessary here. Also, loads from the roots which
previously checked IsOddball would now check IsTrue (etc.), which would
do a load from the roots, which would check IsTrue, ... until the stack
overflowed. Now the root checks do a slow check of the oddball kind for
these oddballs.

Change-Id: I6c0bf1de1fc83ba91831a4a487f638cb4256226b