[shared-struct] Mark shared JS object maps as cannot transition
Maps that can transition are retained weakly by optimized code. There's
a "in retained map" bit on Maps that is set when the map is retained.
Shared object maps are either in shared space or RO space, so setting
this bit is not threadsafe in the case of the former or an access error
in the case of the latter.

It is also more accurate to reflect these as !CanTransition() since
shared JS objects always have fixed layouts.

Bug: v8:12547
Change-Id: Iae1b615c975057342e83910974e5af47fabc3986