[wasm] Force an r-value reference for finalizing methods
There are two methods on the {AsyncCompileJob} which will invalidate the
{this} pointer when we call them.

This CL add rvalue qualifiers to those two methods, which will only
allow to call those methods on r-value ("temporary") objects.
This forces us to use "std::move" on call sites, which clearly indicates
that the object will be invalidated during the method.

R=ahaas@chromium.org

Change-Id: Id6aa53b8627843004d337ece7ec709a9d4620628