[wasm] Merge decrementing finisher count and finishing logic
Before adding a third "finisher" (for validating, see the
"TODO(13447)"), refactor the existing code to merge the logic to finish
a module with the logic to decrement the finisher count. This puts the
code where it belongs and makes the call sites cleaner.

R=ahaas@chromium.org

Bug: v8:13447
Change-Id: I86d64194de9e61d37c879354a34e011606a38046
Cq-Include-Trybots: luci.v8.try:v8_linux64_tsan_dbg,v8_linux64_tsan_rel