[wasm][multi-memory] Support memory.copy
Add support for memory.copy on multiple memories and between different
memories. In contrast to memory.init and memory.fill we need to pass two
memory indexes to the C function that implements memory.copy.

This CL uncovered that we got the order of the two memory index
immediates on memory.copy wrong. I uploaded a spec change to clarify
in the spec that the destination comes first:
https://github.com/WebAssembly/multi-memory/pull/48

R=thibaudm@chromium.org
CC=dlehmann@chromium.org

Bug: v8:13918
Change-Id: I9208b861c7cd286224cc6faa0968f92a1f9da4f4