[heap] Fix marking barrier for shared space
MinorMS generally avoid activating barriers for the shared space.
On page allocation in shared space, page flags were set regrardless of
whether this is a minor or major GC, which could result in the
unexpected case of a shared space page with page flags indicating that
minor incremental marking is in progress (while all existing shared
space pages could only have major incremental marking).

MinorMS also doesn't allocate the shared_heap_worklist_ for clients.
Does for the case of a marking barrier from a client isolate we need to
check not only if marking is active, but whether major marking is
active.
Achieving that (without adding more checks to the barriers) requires
separate major and minor incremental marking page flags.

Bug: chromium:1465120
Change-Id: I9cfd3cf4b35a85d96b587f3f5e05d03248ed4e3c